Add gl3n as dependency and dglsl as optional dependency
Add gl3n 1.3.1 as dependency restricted to 1.3 minor version.
Add dglsl 0.6.0 as a suggested but optional dependency restricted to
minor version 0.6. This is required for further changes to work even
though I have to manually add the files to the source manually due to a
fixed bug in the code that is not fixed in the released version.